+++ Common/src/main/java/net/merchantpug/apugli/registry/action/ApugliItemActions.java
@@ -1,6 +1,7 @@
 package net.merchantpug.apugli.registry.action;
 
 import net.merchantpug.apugli.action.factory.IActionFactory;
+import net.merchantpug.apugli.action.factory.item.CooldownAction;
 import net.merchantpug.apugli.action.factory.item.DamageAction;
 import net.merchantpug.apugli.platform.Services;
 import net.minecraft.util.Tuple;
@@ -11,6 +12,7 @@ import org.apache.commons.lang3.mutable.Mutable;
 public class ApugliItemActions {
     
     public static void registerAll() {
+        register("cooldown", new CooldownAction());
         register("damage", new DamageAction());
     }